Responsive CSS
Adds media queries in site.css which resizes the root font-size depending on the width of the page. Also adds custom properties (CSS variables) to a number of places to make it easier to change sizes in one place.
The canvas is currently resized with a window event instead of in a media query, but the way I ended up doing it I think it could be done in a media query like everything else.