Fixes the game-table and canvas scaling.
Before, the size of the game-table was determined solely the size of its components. The size of 9 cells was calculated plus the in-between space. For some reason (presumably a rounding error) meant that added up to slightly less than the size of the sudoku-board (like 3 or 4 pixels). Now the size of the game-table is set to the same value as the sudoku-board, which makes up for the rounding errors somehow and forces it to be the right size.

The canvas was being scaled with css but on a javascript event. Now it's just resized with media query like everything else so it won't be the wrong size if the page is reloaded.